+++ src/main/java/chatterchicken/data/exception/CCException.java
@@ -1,4 +1,4 @@
-package chatterchicken;
+package chatterchicken.data.exception;
 
 public class CCException extends Exception {
     public CCException(String message) {
