+++ src/main/java/chatterchicken/data/task/Task.java
@@ -1,4 +1,4 @@
-package chatterchicken.task;
+package chatterchicken.data.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
